move CreateAccount to invitation.go
